🔨 Rename mixins module to model_mixins for clarity
