+++ example_replacement_project/user_replacement/models.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import AbstractBaseUser, PermissionsMixin
 from django.utils.translation import ugettext_lazy as _
 
 from improved_user.managers import UserManager
-from improved_user.mixins import DjangoIntegrationMixin, EmailAuthMixin
+from improved_user.model_mixins import DjangoIntegrationMixin, EmailAuthMixin
 
 
 class User(DjangoIntegrationMixin, EmailAuthMixin,
