+++ src/improved_user/model_mixins.py
@@ -1,4 +1,4 @@
-"""Mixin Models to allow for easy creation of User models"""
+"""Mix-in Classes intended for use with Django Models"""
 from django.contrib.auth.models import AbstractBaseUser, PermissionsMixin
 from django.core.mail import send_mail
 from django.db import models
